Renaming assets-folder and partly switching to ES2015
