+++ _assets/javascripts/main.js
@@ -1,0 +1,3 @@
+
+import './lazy-load';
+import './comments';
