+++ _assets/styles/external/normalize.scss/embed/_svg.scss
