+++ _assets/styles/external/normalize.scss/links/_fix-background.scss
