+++ _assets/styles/main.scss
@@ -5,7 +5,6 @@
 
 @import "external/normalize.scss/normalize";
 @import "external/h5bp-pt1";
-@import "../javascripts/vendor/jsonlylightbox/css/lightbox";
 
 @import "syntax-highlighting";
 @import "project/all";
