+++ _assets/styles/project/_lightbox.scss
