+++ _assets/styles/project/_variables.scss
