--- _content/_assets/javascripts/main.js
@@ -1,7 +1,0 @@
-
-//= require vendor/eventie/eventie.js
-//= require vendor/vanilla-pubsub/PubSub.js
-//= require vendor/jsonlylightbox/js/lightbox.js
-//= require lazy-load.js
-//= require lightbox.js
-//= require comments.js
